# contract/openapi.yaml
openapi: 3.1.0
info:
  title: Kestrel AI backend API
  version: 1.0.0
  description: |
    Contract between the Tauri/React frontend and the FastAPI sidecar. This file is the
    source of truth (spec section 2). The backend validates responses against it in tests,
    the TypeScript client is generated from it, and Prism serves it as the mock server.

    Conventions:
    - Every path below carries the `/api/v1` prefix. JSON bodies, UTF-8.
    - Every request carries the per-launch token, either as `Authorization: Bearer <token>`
      or as a `token` query parameter (needed for `<img>` sources and websocket connects).
    - Ids are UUID strings. Timestamps are RFC 3339 strings in UTC.
    - Boxes are in image pixels: `x, y` top-left corner, `w, h` size.
    - Every list endpoint paginates with `limit` and `cursor` and returns `next_cursor`
      (`null` when there is no further page).
    - Every error uses the `Error` envelope `{error: {code, message, details}}`.
    - Long work returns a `Job`; progress arrives over the websocket described in
      `x-websocket` and can be polled through the `jobs` resource.
servers:
  - url: http://127.0.0.1:8765
    description: development backend origin (the packaged app uses a random free port)
  - url: http://127.0.0.1:4010
    description: Prism mock server
security:
  - bearerAuth: []
  - tokenQuery: []
...
paths:
...
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
    get:
      tags: [projects]
      operationId: getProject
      responses:
        "200":
          description: the project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    patch:
      tags: [projects]
      operationId: updateProject
      summary: Update name, pre-annotation model or import defaults (project settings screen).
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProjectUpdate" }
      responses:
        "200":
          description: updated project
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Project" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [projects]
      operationId: forgetProject
      summary: Remove the project from the recent list. The project folder and everything in it stay on disk; "Open folder" brings it back.
      responses:
        "204":
          description: forgotten
        default: { $ref: "#/components/responses/Error" }
...
  /api/v1/projects/{projectId}/design-inspections:
    parameters:
      - $ref: "#/components/parameters/projectId"
    post:
      tags: [surfaces]
      operationId: createDesignInspection
      summary: Read a design file (a `design_import` job, phase `inspect`) - hash it, detect units and CRS, and list its candidate surfaces.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DesignInspectionCreate" }
      responses:
        "202":
          description: inspection created in `inspecting`, job queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DesignInspectionWithJob" }
        "409": { $ref: "#/components/responses/WrongProjectKind" }
        "422":
          description: "the file is a DWG (`code` is `validation_error`, `details.reason` is `\"dwg\"`) or an existing file has an unknown extension (`details.reason` is `\"extension\"`); a missing file is `404 not_found`, never a 422"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/design-inspections/{inspectionId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/inspectionId"
    get:
      tags: [surfaces]
      operationId: getDesignInspection
      responses:
        "200":
          description: the inspection
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DesignInspection" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
    delete:
      tags: [surfaces]
      operationId: deleteDesignInspection
      summary: Cancel a running inspect or preview and delete the inspection folder.
      responses:
        "204": { description: deleted }
        "409":
          description: "a build of this inspection is queued or running (`code` is `job_running`), or the project is not a detection project (`code` is `wrong_project_kind`)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/design-inspections/{inspectionId}/candidates/{candidateId}/thumbnail:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/inspectionId"
      - $ref: "#/components/parameters/candidateId"
    get:
      tags: [surfaces]
      operationId: getDesignCandidateThumbnail
      summary: A 160 px plan view of one candidate.
      responses:
        "200":
          description: thumbnail
          content:
            image/png:
              schema: { type: string, format: binary }
        "204": { description: the candidate has no thumbnail }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/design-inspections/{inspectionId}/previews:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/inspectionId"
    post:
      tags: [surfaces]
      operationId: createDesignPreview
      summary: Place, triangulate and rasterise the chosen candidates onto a coarse preview grid and validate them (a `design_import` job, phase `preview`). It cancels the inspection's running preview first.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DesignImportOptions" }
      responses:
        "202":
          description: preview created in `running`, job queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DesignPreviewWithJob" }
        "409":
          description: "the inspection has not been read yet or reading it failed, or the target cloud surface is building, failed or missing its grid (`code` is `not_ready`); a design surface is being imported from this inspection (`code` is `job_running`); or the project is not a detection project (`code` is `wrong_project_kind`)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "422":
          description: "an unknown or repeated candidate, an unparseable CRS, a LandXML or DEM selection of other than one candidate, a `target_surface_id` that is not a cloud surface (unknown, or not a `cloud_dsm`), or neither a target nor `cell_size_m` (`code` is `validation_error`)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/design-inspections/{inspectionId}/previews/{previewId}:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/inspectionId"
      - $ref: "#/components/parameters/previewId"
    get:
      tags: [surfaces]
      operationId: getDesignPreview
      responses:
        "200":
          description: the preview
          content:
            application/json:
              schema: { $ref: "#/components/schemas/DesignPreview" }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/design-inspections/{inspectionId}/previews/{previewId}/image:
    parameters:
      - $ref: "#/components/parameters/projectId"
      - $ref: "#/components/parameters/inspectionId"
      - $ref: "#/components/parameters/previewId"
    get:
      tags: [surfaces]
      operationId: getDesignPreviewImage
      summary: The preview's plan image.
      responses:
        "200":
          description: preview image
          content:
            image/png:
              schema: { type: string, format: binary }
        "204": { description: the preview is not ready or failed }
        "404": { $ref: "#/components/responses/NotFound" }
        default: { $ref: "#/components/responses/Error" }
  /api/v1/projects/{projectId}/design-surfaces:
    parameters:
      - $ref: "#/components/parameters/projectId"
    post:
      tags: [surfaces]
      operationId: createDesignSurface
      summary: Build the design surface from a ready preview (a `design_import` job, phase `build`); the new `Surface` has `kind` `design`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DesignSurfaceCreate" }
      responses:
        "202":
          description: surface created in `building`, job queued
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SurfaceWithJob" }
        "409":
          description: "the preview is not ready, or the target cloud surface is no longer ready (`code` is `not_ready`); the preview is not the newest, has `block` warnings, or has `warn` warnings without `accept_warnings` (`code` is `conflict`); the design is already being imported (`code` is `job_running`); or the project is not a detection project (`code` is `wrong_project_kind`)"
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        default: { $ref: "#/components/responses/Error" }

  # ------------------------------------------------------------------ volumes (S2 paths 9-17)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: the per-launch token passed to the sidecar in `APP_TOKEN`
    tokenQuery:
      type: apiKey
      in: query
      name: token
      description: same token as a query parameter, for image URLs and the websocket
  parameters:
    projectId:
      name: projectId
      in: path
      required: true
      schema: { type: string }
...
    inspectionId:
      name: inspectionId
      in: path
      required: true
      schema: { type: string }
    previewId:
      name: previewId
      in: path
      required: true
      schema: { type: string }
    candidateId:
      name: candidateId
      in: path
      required: true
      schema: { type: string, pattern: "^c[0-9]{1,6}$" }
  responses:
    Error:
      description: error envelope
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    WrongProjectKind:
      description: "the operation does not belong to this kind of project (`code` is `wrong_project_kind`, details `{kind, allowed}`)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error:
              code: wrong_project_kind
              message: "This is a detection project. Datasets belong in a training project."
              details: { kind: detect, allowed: [train] }
...
    NotFound:
      description: "the project or the named resource does not exist (`code` is `not_found`)"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example:
            error: { code: not_found, message: "run r0000000… not found", details: {} }
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              description: |
                machine-readable: unauthorized, not_found, validation_error, already_exists,
                class_in_use, conflict, not_implemented, provider_error, internal_error,
                wrong_project_kind (409: the operation does not belong to this kind of project;
                details `{kind, allowed}`), library_unavailable (503: the model library could not
                be opened), model_unavailable (409: the library model's weights file is missing),
                unmapped_classes (422: the run's model has classes with no project class and no
                remembered mapping; details `{model_id, unmapped}`), unsupported_point_cloud,
                insufficient_memory and insufficient_disk (422: a point cloud cannot be read or
                admitted), link_needs_coordinates, no_overlap and crs_already_set (422: a point
                cloud patch), invalid_epsg (422: an EPSG code pyproj does not know),
                wrong_point_count, needs_projected_crs, vertical_span_too_small and
                measurement_limit (422: a point-cloud measurement), octree_missing (404: a ready
                cloud's display-copy file is missing on disk), grid_too_large (422: a surface
                grid over the cell ceiling), unsupported_crs, source_missing and
                invalid_build_request (422: a surface build), no_coordinates (422: an ortho
                underlay without a CRS), invalid_geometry and invalid_base (422: a volume
                measurement's polygon or base), job_running (409: the resource's job is queued
                or running), not_ready (409: the resource has not finished importing),
                range_not_satisfiable (416)
            message: { type: string }
            details:
              type: object
              additionalProperties: true
      example:
        error: { code: not_found, message: "project 7f1c… not found", details: {} }
...
    ClassDef:
      type: object
      required: [id, name, colour, hotkey, order]
      properties:
        id: { type: string }
        name: { type: string }
        colour: { type: string }
        hotkey: { type: [string, "null"] }
        order: { type: integer }
      example: { id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", hotkey: "1", order: 0 }
    ImportSettings:
      type: object
      description: preparation settings (spec section 5); missing fields take the project defaults
      properties:
        max_side: { type: integer, minimum: 512, maximum: 12000, default: 4000 }
        quality: { type: integer, minimum: 50, maximum: 100, default: 95 }
        dedupe_threshold: { type: integer, minimum: 0, maximum: 32, default: 4, description: max phash Hamming distance for a duplicate }
        group_regex:
          type: string
          default: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)"
          description: the named group `flight` is the group key (flights span cameras); fall back to a 250 m tile, then the source site
      example: { max_side: 4000, quality: 95, dedupe_threshold: 4, group_regex: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)" }
    ProjectKind:
      type: string
      enum: [train, detect]
      description: "`train` labels images, builds datasets and trains models; `detect` runs library models over sources. Set at creation, never changed."
      example: train
    Project:
      type: object
      required: [id, name, folder, kind, classes, preannotation_model_id, import_defaults, schema_version, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        folder: { type: string, description: absolute path of the project folder }
        kind: { $ref: "#/components/schemas/ProjectKind" }
        classes:
          type: array
          items: { $ref: "#/components/schemas/ClassDef" }
        preannotation_model_id: { type: [string, "null"], description: a library model id }
        import_defaults: { $ref: "#/components/schemas/ImportSettings" }
        schema_version: { type: integer }
        created_at: { type: string, format: date-time }
      example:
        id: "7f1c2e3a-1111-4000-8000-000000000001"
        name: Ahmadia
        folder: "E:\\Projects\\Ahmadia"
        kind: train
        classes:
          - { id: "c1a2b3c4-0000-4000-8000-000000000001", name: excavator, colour: "#f97316", hotkey: "1", order: 0 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000002", name: wheel_loader, colour: "#eab308", hotkey: "2", order: 1 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000003", name: bulldozer, colour: "#22c55e", hotkey: "3", order: 2 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000004", name: dump_truck, colour: "#06b6d4", hotkey: "4", order: 3 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000005", name: crane, colour: "#3b82f6", hotkey: "5", order: 4 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000006", name: concrete_mixer, colour: "#a855f7", hotkey: "6", order: 5 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000007", name: roller, colour: "#ec4899", hotkey: "7", order: 6 }
          - { id: "c1a2b3c4-0000-4000-8000-000000000008", name: backhoe, colour: "#ef4444", hotkey: "8", order: 7 }
        preannotation_model_id: "m0000000-2222-4000-8000-000000000001"
        import_defaults: { max_side: 4000, quality: 95, dedupe_threshold: 4, group_regex: "^(?P<camera>[A-Za-z0-9-]+)_(?P<flight>\\d+)_(?P<frame>\\d+)" }
        schema_version: 1
        created_at: "2026-09-17T10:00:00Z"
...
    ProjectUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        preannotation_model_id: { type: [string, "null"], description: a library model id }
        import_defaults: { $ref: "#/components/schemas/ImportSettings" }
      example: { name: Ahmadia, preannotation_model_id: "m0000000-2222-4000-8000-000000000001" }
...
    SurfaceWithJob:
      type: object
      required: [surface, job]
      properties:
        surface: { $ref: "#/components/schemas/Surface" }
        job: { $ref: "#/components/schemas/Job" }
...
    DesignFormat: { type: string, enum: [geotiff, landxml, dxf] }
    LinearUnit: { type: string, enum: [millimetre, centimetre, metre, international_foot, us_survey_foot] }
...
    DesignInspectionCreate:
      type: object
      required: [path]
      properties:
        path: { type: string, minLength: 1, description: absolute path of a .tif/.tiff, .xml/.landxml or .dxf file }
...
    DesignInspection:
      type: object
      required: [id, state, error, job_id, path, format, file_size, sha256, detected, candidates, default_target_surface_id, created_at]
      properties:
        id: { type: string }
        state: { type: string, enum: [inspecting, ready, failed] }
        error: { type: [string, "null"] }
        job_id: { type: string }
        path: { type: string }
        format: { $ref: "#/components/schemas/DesignFormat" }
        file_size: { type: integer }
        sha256: { type: [string, "null"] }
        detected: { oneOf: [{ $ref: "#/components/schemas/DesignDetected" }, { type: "null" }] }
        candidates: { type: array, items: { $ref: "#/components/schemas/DesignCandidate" } }
        default_target_surface_id: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
    DesignInspectionWithJob:
      type: object
      required: [inspection, job]
      properties:
        inspection: { $ref: "#/components/schemas/DesignInspection" }
        job: { $ref: "#/components/schemas/Job" }
    DesignImportOptions:
      type: object
      required: [candidate_ids, source_crs, horizontal_unit, vertical_unit]
      properties:
        candidate_ids: { type: array, minItems: 1, items: { type: string, pattern: "^c[0-9]{1,6}$" } }
        source_crs: { type: string, minLength: 1, description: "'EPSG:<code>' or WKT; parsed with pyproj.CRS.from_user_input" }
        horizontal_unit: { $ref: "#/components/schemas/LinearUnit" }
        vertical_unit: { $ref: "#/components/schemas/LinearUnit" }
        swap_xy: { type: boolean, description: "false when absent" }
        target_surface_id: { type: [string, "null"], description: "null when absent. A ready cloud_dsm surface; the output adopts its CRS and cell size on the aligned lattice (same_lattice with it)" }
        cell_size_m: { type: [number, "null"], exclusiveMinimum: 0, description: "null when absent; required when target_surface_id is null; ignored otherwise; default in the UI 0.25" }
        max_edge_m: { type: [number, "null"], minimum: 0, description: "null when absent. Points geometry only; null = automatic, 0 = off" }
...
    DesignZCheck:
      type: object
      required: [median_dz_m, p05_dz_m, p95_dz_m, n_samples, design_z_min_m, design_z_max_m]
      properties:
        median_dz_m: { type: number }
        p05_dz_m: { type: number }
        p95_dz_m: { type: number }
        n_samples: { type: integer }
        design_z_min_m: { type: number }
        design_z_max_m: { type: number }
    DesignPreview:
      type: object
      required: [id, inspection_id, state, error, job_id, options, output, triangle_count, overlap_fraction, target_covered_fraction, design_area_m2, z_check, warnings, suggestions, created_at]
      properties:
        id: { type: string }
        inspection_id: { type: string }
        state: { type: string, enum: [running, ready, failed] }
        error: { type: [string, "null"] }
        job_id: { type: string }
        options: { $ref: "#/components/schemas/DesignImportOptions" }
        output: { oneOf: [{ $ref: "#/components/schemas/DesignPreviewOutput" }, { type: "null" }] }
        triangle_count: { type: [integer, "null"] }
        overlap_fraction: { type: [number, "null"] }
        target_covered_fraction: { type: [number, "null"] }
        design_area_m2: { type: [number, "null"] }
        z_check: { oneOf: [{ $ref: "#/components/schemas/DesignZCheck" }, { type: "null" }] }
        warnings: { type: array, items: { $ref: "#/components/schemas/DesignWarning" } }
        suggestions: { type: array, items: { $ref: "#/components/schemas/DesignSuggestion" } }
        created_at: { type: string, format: date-time }
    DesignPreviewWithJob:
      type: object
      required: [preview, job]
      properties:
        preview: { $ref: "#/components/schemas/DesignPreview" }
        job: { $ref: "#/components/schemas/Job" }
    DesignSurfaceCreate:
      type: object
      required: [inspection_id, preview_id]
      properties:
        inspection_id: { type: string }
        preview_id: { type: string }
        name: { type: string, minLength: 1, maxLength: 200 }
        accept_warnings: { type: boolean, description: "false when absent" }
...